requirements.txt
Django==3.1.5
django_redis==4.12.1
djangorestframework==3.12.4
django-filter==2.4.0
pillow==8.2.0
djangorestframework-jwt==1.11.0
